Add a way to reopen a gist without using cache
This commit close #31. It add a way to reopen an opend gist without
using cache.
vim-gista cache a gist when it is opened to accelerate the response. But
sometime user want to re-download the gist for reverting or so on.

If user want to re-download the gist, there are several way

1.  Use upper case open keys (E, S, V) on gist:list buffer to re-open
    the selected gist
2.  Use 'open!' command on Unite buffer to re-open the selected gist
3.  Use `:Gista -o` command with a bang (!) like `:Gista! -o` on the
    buffer which connected to a gist